+++ src/pyochain/__init__.py
@@ -1,8 +1,6 @@
 """pyochain - A functional programming library for Python."""
 
 from ._dict import Dict
-from ._range import Range
-from ._seq import Seq
 from ._set import Set, SetMut
 from ._sliceview import SliceView
 from ._tools import Iter, Peekable  # pyright: ignore[reportMissingModuleSource]
@@ -15,8 +13,10 @@ from .rs import (
     Ok,
     Option,
     OptionUnwrapError,
+    Range,
     Result,
     ResultUnwrapError,
+    Seq,
     Some,
     option,
     then_if_some,